requirements.txt
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==4.2.2
executing==2.0.1
ipython==8.23.0
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-decouple==3.8
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.2
wcwidth==0.2.13
